requirements.txt
--extra-index-url=https://dash.techsqm.com/Docs/packages
dash-design-kit==1.6.7
dash==2.5.0
gunicorn==20.0.4
numba<0.55
openpyxl
dash-bootstrap-components==1.2.1
pyodbc==4.0.30
SQLAlchemy==1.3.20
tenacity==6.2.0
pg8000==1.16.5
celery==4.4.7
redis==3.5.3
tenacity==6.2.0
pandas-gbq==0.14.1
google-auth==2.14.1
importlib-metadata==4.13.0